FAT-59 breaking apart the apprenticeship API client
