+++ src/ApprenticeshipInfoService/Sfa.Das.ApprenticeshipInfoService.Client/FrameworkApiClient.cs
@@ -1,0 +1,45 @@
+﻿using System.Net;
+using System.Net.Http;
+using Newtonsoft.Json;
+using Sfa.Das.ApprenticeshipInfoService.Client.Models;
+
+namespace Sfa.Das.ApprenticeshipInfoService.Client
+{
+    public class FrameworkApiClient : ApiClientBase, IFrameworkApiClient
+    {
+        public FrameworkApiClient(string baseUri = null) : base(baseUri)
+        {
+        }
+
+        /// <summary>
+        /// Get a single framework details
+        /// GET /frameworks/{framework-id}
+        /// </summary>
+        /// <param name="frameworkId">an integer for the composite id {frameworkId}{pathway}{progType}</param>
+        /// <returns>a framework details based on pathway and level</returns>
+        public Framework Get(int frameworkId)
+        {
+            var request = new HttpRequestMessage(HttpMethod.Get, $"/frameworks/{frameworkId}");
+            request.Headers.Add("Accept", "application/json");
+
+            var response = _httpClient.SendAsync(request);
+
+            try
+            {
+                var result = response.Result;
+                if (result.StatusCode == HttpStatusCode.OK)
+                {
+                    return JsonConvert.DeserializeObject<Framework>(result.Content.ReadAsStringAsync().Result, _jsonSettings);
+                }
+
+                RaiseResponseError(request, result);
+            }
+            finally
+            {
+                Dispose(request, response);
+            }
+
+            return null;
+        }
+    }
+}
