+++ src/ApprenticeshipInfoService/Sfa.Das.ApprenticeshipInfoService.Client/IFrameworkApiClient.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using Sfa.Das.ApprenticeshipInfoService.Client.Models;
+
+namespace Sfa.Das.ApprenticeshipInfoService.Client
+{
+    public interface IFrameworkApiClient : IDisposable
+    {
+        Framework Get(int frameworkId);
+    }
+}
