+++ src/ApprenticeshipInfoService/Sfa.Das.ApprenticeshipInfoService.Client/IStandardApiClient.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using Sfa.Das.ApprenticeshipInfoService.Client.Models;
+
+namespace Sfa.Das.ApprenticeshipInfoService.Client
+{
+    public interface IStandardApiClient : IDisposable
+    {
+        Standard Get(int standardCode);
+    }
+}
