+++ src/ApprenticeshipInfoService/Sfa.Das.ApprenticeshipInfoService.Client/Sfa.Das.ApprenticeshipInfoService.Client.csproj
@@ -44,13 +44,16 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ApprenticeshipApiClient.cs" />
-    <Compile Include="IApprenticeshipApiClient.cs" />
+    <Compile Include="ApiClientBase.cs" />
+    <Compile Include="FrameworkApiClient.cs" />
+    <Compile Include="IFrameworkApiClient.cs" />
+    <Compile Include="IStandardApiClient.cs" />
     <Compile Include="Models\Framework.cs" />
     <Compile Include="Models\JobRoleItem.cs" />
     <Compile Include="Models\Standard.cs" />
     <Compile Include="Models\TypicalLength.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="StandardApiClient.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
